+++ ArchersGame/Lightning/Lightning.hpp
@@ -24,8 +24,10 @@ public:
     void setWindow(sf::RenderWindow* gameWindow);
     void setSize(int width, int height);
     void updateFrame(double time);
+
     void change_volume_lightning(float volume);
 
-    int getLightningRate();
-    void setLightningRate(int rate);
+    float getLightningRate();
+    void setLightningRate(float rate);
+
 };
